优化common_task_template字段类型

+++ assets/db_table/common_button.sql
@@ -66,10 +66,10 @@ create index button_end_time_index on common_task (`end_time`);
 
 create table common_task_template
 (
-    `id`          int unsigned auto_increment   not null primary key comment '任务模板id',
-    `period_type` tinyint unsigned default 0    not null comment '任务周期：0=无周期 1=自然日 2=自然周(第一天是周日) 3=自然周(第一天是周一) 4=自然月',
-    `task_type`   tinyint unsigned default 0    not null comment '任务类型：1=跳转任务 2=签到',
-    `extend`      varchar(8192)    default '{}' not null comment '扩展数据, 一般用于存放任务模板数据的参数, 这些参数是开发者决定的, 比如第三方任务的id和secret',
+    `id`          int unsigned auto_increment    not null primary key comment '任务模板id',
+    `period_type` smallint unsigned default 0    not null comment '任务周期：0=无周期 1=自然日 2=自然周(第一天是周日) 3=自然周(第一天是周一) 4=自然月',
+    `task_type`   smallint unsigned default 0    not null comment '任务类型：1=跳转任务 2=签到',
+    `extend`      varchar(8192)     default '{}' not null comment '扩展数据, 一般用于存放任务模板数据的参数, 这些参数是开发者决定的, 比如第三方任务的id和secret',
     -- 以下字段无业务逻辑, 或者是透传到客户端字段
-    `remark`      varchar(256)     default ''   not null comment '备注'
+    `remark`      varchar(256)      default ''   not null comment '备注'
 ) comment ='通用任务模板'
